+++ lib/idempotent-request.rb
@@ -1,5 +1,7 @@
 require 'oj'
+require 'yaml'
 require 'idempotent-request/request'
 require 'idempotent-request/request_manager'
 require 'idempotent-request/redis_storage'
+require 'idempotent-request/policy'
 require 'idempotent-request/middleware'