+++ lib/idempotent-request/middleware.rb
@@ -2,10 +2,11 @@ module IdempotentRequest
   class Middleware
     def initialize(app, config = {})
       @app = app
-      @config = config
-      @policy = config.fetch(:policy)
+      @config = config.merge(load_config(config.fetch(:config_file, 'config/idempotent.yml')))
+      @policy = @config.fetch(:policy, IdempotentRequest::Policy)
+      @concurrent_response_status = @config.fetch(:concurrent_response_status, 429)
+      @replayed_response_header = @config.fetch(:replayed_response_header, 'Idempotency-Replayed')
       @notifier = ActiveSupport::Notifications if defined?(ActiveSupport::Notifications)
-      @conflict_response_status = config.fetch(:conflict_response_status, 429)
     end
 
     def call(env)
@@ -15,9 +16,8 @@ module IdempotentRequest
 
     def process(env)
       set_request(env)
-      request.env['idempotent.request'] = {}
       return app.call(request.env) unless process?
-      request.env['idempotent.request']['key'] = request.key
+      request.env['idempotent.request'] = { key: request.key }
       response = read_idempotent_request || write_idempotent_request || concurrent_request_response
       instrument(request)
       response
@@ -26,29 +26,49 @@ module IdempotentRequest
     private
 
     def storage
-      @storage ||= RequestManager.new(request, config)
+      @storage ||= RequestManager.new(request, config.merge(expire_time: expire_time_for_route))
+    end
+
+    def expire_time_for_route
+      policy_instance = policy.new(request, config)
+      policy_instance.respond_to?(:expire_time_for_route) ? policy_instance.expire_time_for_route : nil
     end
 
     def read_idempotent_request
-      request.env['idempotent.request']['read'] = storage.read
+      result = storage.read
+      return unless result
+
+      status, headers, response = result
+      headers.merge!(@replayed_response_header => true)
+      request.env['idempotent.request']['read'] = [status, headers, response]
     end
 
     def write_idempotent_request
-      return unless storage.lock
+      # Only consider 'false' lock result as key existed, and treat as concurrent request
+      # Consider 'true', nil, or other values as key not existed, and continue as normal request
+      return if storage.lock == false
+
       begin
         result = app.call(request.env)
-        request.env['idempotent.request']['write'] = result
         storage.write(*result)
+        request.env['idempotent.request']['write'] = result
       ensure
         request.env['idempotent.request']['unlocked'] = storage.unlock
-        result
       end
+
+      result
     end
 
     def concurrent_request_response
-      status = @conflict_response_status
+      status = @concurrent_response_status
       headers = { 'Content-Type' => 'application/json' }
-      body = [ Oj.dump('error' => 'Concurrent requests detected') ]
+      body = [
+        Oj.dump("error" => {
+          "type" => "TooManyRequests",
+          "message" => "Concurrent requests detected",
+          "code" => "too_many_requests"
+        })
+      ]
       request.env['idempotent.request']['concurrent_request_response'] = true
       Rack::Response.new(body, status, headers).finish
     end
@@ -61,7 +81,7 @@ module IdempotentRequest
 
     def should_be_idempotent?
       return false unless policy
-      policy.new(request).should?
+      policy.new(request, config).should?
     end
 
     def instrument(request)
@@ -71,6 +91,29 @@ module IdempotentRequest
     def set_request(env)
       @env = env
       @request ||= Request.new(env, config)
+    end
+
+    def load_config(config_file)
+      return {} unless File.exist?(config_file)
+      
+      config = YAML.load_file(config_file) || {}
+
+      if config.respond_to? :deep_symbolize_keys!
+        config.deep_symbolize_keys!
+      else
+        symbolize_keys_deep!(config)
+      end
+
+      environment = ENV["APP_ENV"] || ENV["RAILS_ENV"] || ENV["RACK_ENV"] || "default"
+      config[environment.to_sym] || {}
+    end
+    
+    def symbolize_keys_deep!(hash)
+      hash.keys.each do |k|
+        symkey = k.respond_to?(:to_sym) ? k.to_sym : k
+        hash[symkey] = hash.delete k
+        symbolize_keys_deep! hash[symkey] if hash[symkey].is_a? Hash
+      end
     end
   end
 end