+++ lib/idempotent-request/request_manager.rb
@@ -1,15 +1,16 @@
 module IdempotentRequest
   class RequestManager
-    attr_reader :request, :storage
+    attr_reader :request, :storage, :expire_time
 
     def initialize(request, config)
       @request = request
       @storage = config.fetch(:storage)
+      @expire_time = config[:expire_time]
       @callback = config[:callback]
     end
 
     def lock
-      storage.lock(key)
+      storage.lock(key, expire_time)
     end
 
     def unlock
@@ -20,7 +21,7 @@ module IdempotentRequest
       status, headers, response = parse_data(storage.read(key)).values
 
       return unless status
-      run_callback(:detected, key: request.key)
+      run_callback(:detected, key: key)
       [status, headers, response]
     end
 
@@ -29,7 +30,7 @@ module IdempotentRequest
       response = response.body if response.respond_to?(:body)
 
       if (200..226).cover?(status)
-        storage.write(key, payload(status, headers, response))
+        storage.write(key, payload(status, headers, response), expire_time)
       end
 
       data